[refactor] : producer spread 로직 분리
